prompt_caching/requirements.txt
crewai>=0.11.0
streamlit>=1.32.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0 